Feat: 요약 및 문제 생성 API 구현 (#34)
* Init: 개발환경 세팅 (#2)

* Init: 프로젝트 기본설정 세팅

- 프로젝트 생성
- .gitignore설정
- 프로젝트 의존성 추가
- application.yml 설정파일 구성

* Init: 프로젝트 기본 구조 및 공통 컴포넌트 설정

- 공통 설정 클래스 추가 (JPA, QueryDSL, Swagger, Web)
- 공통 도메인 엔티티 (RootEntity) 정의
- 예외 처리 관련 클래스 및 타입 구현
- JSON 변환을 위한 AttributeConverter 추가
- 유틸리티 클래스 (Math) 추가

* Feat: Member 기능 구현 (#15)

* Feat: Member 도메인 구현

* Feat: Kakao 로그인 Client로직 구현

- 안드로이드에서 전달 된 이미 검증이 된 Kakao accesstoken으로 Member조회

* Test: Kakao 로그인 fetchMember 단위테스트 작성

* Feat: JWT 관련 기능 구현

- Interceptor에서 특정 path를 제외하고 토큰이 유효한지 검증
- @Auth 어노테이션이 붙어있으면서 Long타입인 파라미터인 경우 ArgumentResolver에서 토큰검증 및 memberId를 반환해주도록 구현

* Feat: Member API 구현

* Style: 불필요한 import 제거

* Test: 하드코딩 된 값 변수추출

* Refactor: 불필요한 토큰 검증 제거

- Interceptor에서 검증된 토큰으로 memberId를 추출해 ArgumenResolver에서 그 memberId를 사용할 수 있도록 개선

* Feat: Member 기능 구현 (#20)

* Feat: Member 도메인 구현

* Feat: Kakao 로그인 Client로직 구현

- 안드로이드에서 전달 된 이미 검증이 된 Kakao accesstoken으로 Member조회

* Test: Kakao 로그인 fetchMember 단위테스트 작성

* Feat: JWT 관련 기능 구현

- Interceptor에서 특정 path를 제외하고 토큰이 유효한지 검증
- @Auth 어노테이션이 붙어있으면서 Long타입인 파라미터인 경우 ArgumentResolver에서 토큰검증 및 memberId를 반환해주도록 구현

* Feat: Member API 구현

* Style: 불필요한 import 제거

* Test: 하드코딩 된 값 변수추출

* Refactor: 불필요한 토큰 검증 제거

- Interceptor에서 검증된 토큰으로 memberId를 추출해 ArgumenResolver에서 그 memberId를 사용할 수 있도록 개선

* Feat: 폴더, 문서 엔티티 생성, 연결

* Init: 프로젝트 기본설정 세팅

- 프로젝트 생성
- .gitignore설정
- 프로젝트 의존성 추가
- application.yml 설정파일 구성

* Init: 프로젝트 기본 구조 및 공통 컴포넌트 설정

- 공통 설정 클래스 추가 (JPA, QueryDSL, Swagger, Web)
- 공통 도메인 엔티티 (RootEntity) 정의
- 예외 처리 관련 클래스 및 타입 구현
- JSON 변환을 위한 AttributeConverter 추가
- 유틸리티 클래스 (Math) 추가

* Chore: Folder 도메인 폴더 구조 셋업

폴더 구조 셋업 작업

* Feat: Folder 도메인의 엔티티 생성

엔티티 생성자, 부모-자식간 연결로직 생성

* refactor: 자기 참조 관계 설정 수정

기존, 다대일 양방향 관계에서 다대일 단방향 관계로 설정하고, 삭제 등의 이슈 발생시 Service 계층에서 함수의 재귀사용을 통해 삭제할 예정

* Chore: Document 도메인 폴더 구조 셋업

폴더 구조 셋업 및 엔티티 생성

* Feat: Lombok 라이브러리 활용하여 기본 생성자 생성

기본 생성자 생성 lombok 라이브러리 활용하여 대체

* chore: name 필드의 length 50으로 설정

name 필드 (Document, Folder) 의 length = 50 으로 설정

* chore: Domain 계층의 Repository가 QueryRepository 상속받도록 함

상속 작업 수행

* chore: Member 도메인 매핑 작업 수행

Member 도메인 매핑 작업 수행

---------


* Feat: Summary, Problem, AITask 엔티티 생성 (#26)

* Feat: Summary 도메인 구현

* Feat: Problem 도메인 구현

* Feat: AI Task 도메인 구현

* Feat: 요약 및 문제 생성 API 틀 작성

- 필요한 DTO, Command 클래스 작성
- 내부 구현 흐름 파악용 임시 메서드 작성

* Rename: AITask 패키지 이름 변경

* Chore: 불필요한 어노테이션 삭제

* Comment: 임시 반환값에 대한 주석 추가

* Chore: 복합어 엔티티 테이블 이름 명시

- AITask 엔티티 @Table 어노테이션 추가

* Feat: Post  기능 구현

* Feat: Post 기본 코드 구현
- 깃허브가 꼬여서 이전에 작성한거 다시 가지고 왔습니다..

* Fix: Post - Member Id 외래키 매핑

* Fix: Post - @NotNull 적용

* Feat: Comment 기능구현

* Feat : Comment 기본코드 가져오기
- 이전에 작성한 깃이 꼬여서 다시 가져왔습니다..

* Fix : Comment - Member Id 연관관계 매핑

* Fix: Comment - @NotNull 적용

* Feat: 요약 및 문제 생성 API 구현 (#33)

* Style: formatting통일 (#30)

* Feat: 요약 및 문제 생성 API 구현 (#32)

* Rename: AI Task 도메인 이름 변경

- llm 으로 변경

* Refactor: LLM 도메인 애플리케이션 계층과 프레젠테이션 계층 응답 분리

* Feat: LLM 작업 진행 상태 확인 기능 구현

* Feat: 요약 및 문제 결과 조회 기능 구현

* Feat: 요약 및 문제 생성 기능 구현

- AI 서버와 통신하는 부분 제외하고 기능 구현
- 임시 UUID 를 통해 task 저장

* Feat: LLM 서버 콜백 기능 구현

- LLM 서버가 API 콜을 통해 페이지별 요약 및 문제 내용 전달
- task id 를 통해 조회하여 요약 내용과 문제 내용 업데이트

* Refactor: 변수 이름 변경

- 필드명 카멜케이스로 변경

* Refactor: 통일성 없는 부분 수정

- 필드명 변경
- 변수 추출

* Refactor: 예외 종류, 메서드 네이밍 변경

- LLMQueryService 예외 타입 변경
- SummaryAndProblemUpdateResponse 메서드 네이밍 변경

* Refactor: LLMQueryService 응답과 LLMController 응답 분리

---------


---------

+++ src/main/java/notai/member/application/MemberService.java
@@ -16,6 +16,7 @@ public class MemberService {
     public Long login(Member member) {
         return memberRepository.findByOauthId(member.getOauthId())
                 .orElseGet(() -> memberRepository.save(member))
-                .getId();
+                .getId()
+                .longValue();
     }
 }
